requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
Django==3.1
django-environ==0.4.5
imageio==2.9.0
joblib==0.17.0
kiwisolver==1.2.0
matplotlib==3.3.2
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.3
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tifffile==2020.10.1
